PopupBehavior: refactor popup request handling
Now handle redirect requests by a newly introduced PopupBehavior,
which is similar in spirit to the Nimiq Hub's PopupBehavior but
different in implementation because TEN31 Pass does not use @nimiq/rpc.

Redirect related code is now neatly encapsulated in this class making
it easier to extend it with more features, like the popup overlay which
is comming up in the next commit.